+++ klini.cu
@@ -10,7 +10,7 @@ PROCTYPE int StaticWhite(uint32_t w, uint32_t b, uint32_t d){
 	uint32_t idx;
         Pack(&busy,&iwhite,&idamka,w,b,d);
 		idx = blist_get(busy);
-        switch(twobit_get(array + ((iwhite<<RANK)|idamka) * CNK/4, idx)){
+        switch(twobit_get(array + (uint64_t)((iwhite<<RANK)|idamka) * CNK/4, idx)){
         case 3 : // Cimus ZZ
 		return 0;
         case 0 :
@@ -38,7 +38,7 @@ KERNEL
 #endif
     unsigned busy;
     unsigned idx;
-    unsigned char * job = array + ij * CNK /4;
+    unsigned char * job = array + (uint64_t)ij * CNK /4;
     for(idx=0,busy=ALLONE(RANK);_popc(busy)==RANK;idx++,busy = _permut(busy))
 	if(twobit_get(job,idx)==0) {
 	        uint32_t w,b,d;
